// src/app/Models/BordereaudeClotureDecaisse.model.ts
export interface BordereauDeClotureDecaisse {
  bordereau_id?: number; 
  caisse_vendeur_id: number;
  date_comptable: string; 
  Billets?: number; 
  Valeur_Unitaire_Billet_10000?: number; 
  Billets_10000_nombre?: number;
  Billets_10000_somme?: number;
  Valeur_Unitaire_Billet_5000?: number; 
  Billets_5000_nombre?: number;
  Billets_5000_somme?: number;
  Valeur_Unitaire_Billet_2000?: number; 
  Billets_2000_nombre?: number;
  Billets_2000_somme?: number;
  Valeur_Unitaire_Billet_1000?: number;
  Billets_1000_nombre?: number;
  Billets_1000_somme?: number;
  Valeur_Unitaire_Billet_500?: number; 
  Billets_500_nombre?: number;
  Billets_500_somme?: number;
  Pieces?: number; 
  Valeur_Unitaire_Piece_500?: number; 
  Pieces_500_nombre?: number;
  Pieces_500_somme?: number;
  Valeur_Unitaire_Piece_100?: number; 
  Pieces_100_nombre?: number;
  Pieces_100_somme?: number;
  Valeur_Unitaire_Piece_50?: number; 
  Pieces_50_nombre?: number;
  Pieces_50_somme?: number;
  Valeur_Unitaire_Piece_25?: number; 
  Pieces_25_nombre?: number;
  Pieces_25_somme?: number;
}